+++ Albatross/Yokai Watch/Games/YW2/Logic/Charaparam.cs
@@ -31,7 +31,7 @@ namespace Albatross.Yokai_Watch.Games.YW2.Logic
         public new float AttributeDamageEarth { get => base.AttributeDamageEarth; set => base.AttributeDamageEarth = value; }
         public new float AttributeDamageLigthning { get => base.AttributeDamageLigthning; set => base.AttributeDamageLigthning = value; }
         public new float AttributeDamageWater { get => base.AttributeDamageWater; set => base.AttributeDamageWater = value; }
-        public new int AttributeDamageWind { get => base.AttributeDamageWind; set => base.AttributeDamageWind = value; }
+        public new float AttributeDamageWind { get => base.AttributeDamageWind; set => base.AttributeDamageWind = value; }
         public new int SoultimateHash { get => base.SoultimateHash; set => base.SoultimateHash = value; }
         public int Unk7 { get; set; }
         public int Unk8 { get; set; }
